remove tfutil
